+++ src/endstone/python/level.cpp
@@ -22,6 +22,28 @@ Location create_location(const Nullable<Dimension> &dimension, float x, float y,
 {
     return {dimension, x, y, z, pitch, yaw};
 }
+
+py::list get_recipes(Level &level)
+{
+    py::list result;
+    for (auto &recipe : level.getRecipes()) {
+        switch (recipe->getType()) {
+        case RecipeType::Smithing:
+            result.append(py::cast(std::unique_ptr<SmithingRecipe>(static_cast<SmithingRecipe *>(recipe.release()))));
+            break;
+        case RecipeType::Shaped:
+            result.append(py::cast(std::unique_ptr<ShapedRecipe>(static_cast<ShapedRecipe *>(recipe.release()))));
+            break;
+        case RecipeType::Shapeless:
+            result.append(py::cast(std::unique_ptr<ShapelessRecipe>(static_cast<ShapelessRecipe *>(recipe.release()))));
+            break;
+        case RecipeType::Multi:
+            result.append(py::cast(std::unique_ptr<MultiRecipe>(static_cast<MultiRecipe *>(recipe.release()))));
+            break;
+        }
+    }
+    return result;
+}
 }  // namespace
 
 void init_level(py::module_ &m, py::classh<Level> &level, py::classh<Dimension> &dimension,
@@ -161,8 +183,7 @@ void init_level(py::module_ &m, py::classh<Level> &level, py::classh<Dimension> 
 )doc")
         .def_property_readonly("actors", &Dimension::getActors,
                                "A list of all actors currently residing in this dimension.")
-        .def_property_readonly("mobs", &Dimension::getMobs,
-                               "A list of all mobs currently residing in this dimension.")
+        .def_property_readonly("mobs", &Dimension::getMobs, "A list of all mobs currently residing in this dimension.")
         .def_property_readonly("players", &Dimension::getPlayers,
                                "A list of all players currently residing in this dimension.");
 
@@ -182,6 +203,9 @@ void init_level(py::module_ &m, py::classh<Level> &level, py::classh<Dimension> 
         .def_property_readonly("actors", &Level::getActors, "A list of all actors currently residing in this level.")
         .def_property("time", &Level::getTime, &Level::setTime, "The relative in-game time of this level.")
         .def_property_readonly("dimensions", &Level::getDimensions, "A list of all dimensions within this level.")
+        .def_property_readonly("recipes", &get_recipes, R"doc(
+    A snapshot of the crafting recipes currently registered by the level.
+)doc")
         .def("get_dimension", &Level::getDimension, py::arg("id"), R"doc(
     Gets the dimension with the given id.